routerbot/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-groq
langgraph
groq
sentence-transformers
faiss-cpu
numpy
pydantic
transformers
torch
pypdf